refactor(THU-601): extract shared bridge Step and link install-command constants
